does not work but it addresses some issues on 32 bit machines for compressed arrays

+++ src/realm/array.cpp
@@ -269,7 +269,7 @@ void Array::init_from_mem(MemRef mem) noexcept
         // encoder knows which format we are compressed into, width and size are read accordingly with the format of
         // the header
         m_size = m_encoder.size();
-        m_width = m_encoder.width();
+        m_width = static_cast<uint_least8_t>(m_encoder.width());
         // we need to compute lower and upper bound, these are useful during Array::find and in general in every query
         // related optimisation.
         if (m_width) {
